# docs/swagger.json
{
    "swagger":"2.0",
    "info":{
        "description":"Template Store::: APIs to get, post and render templates.",
        "version":"1.0.0",
        "title":"Template Store"
    },
    "host":"127.0.0.1:8000",
    "basePath":"/templatestore/api/v1",
    "schemes":[
        "https",
        "http"
    ],
    "paths":{
        "/template":{
            "post":{
                "tags":[
                    "Add Template"
                ],
                "summary":"Add a new template to the template store",
                "description":"",
                "operationId":"post_template",
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"Template properties to push to template store",
                        "required":true,
                        "schema":{
                            "type":"object",
                            "properties":{
                                "name":{
                                    "type":"string",
                                    "maxLength":1000
                                },
                                "sub_template": {
                                    "type":"array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "sub_type": {
                                                "type": "string"
                                            },
                                            "data": {
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        }
                                    }
                                },
                                "type":{
                                    "type":"string",
                                    "description":"allowed types should be preconfigured in TemplateConfig table"
                                },
                                "attributes":{
                                    "type":"object",
                                    "additionalProperties":true
                                },
                                "sample_context_data":{
                                    "type":"object",
                                    "additionalProperties":true
                                }
                            }
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "name":{
                                    "type":"string",
                                    "maxLength":1000
                                },
                                "version":{
                                    "type":"string",
                                    "pattern":"^\\d+\\.\\d+$"
                                },
                                "default":{
                                    "type":"boolean"
                                },
                                "attributes":{
                                    "type":"object",
                                    "additionalProperties":true
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Bad Request",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "error":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/template/{name}/{version}":{
            "get":{
                "tags":[
                    "Get Template Details"
                ],
                "summary":"Fetches template details",
                "description":"",
                "operationId":"get_template_details",
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"name",
                        "in":"path",
                        "description":"name of the template",
                        "required":true,
                        "type":"string"
                    },
                    {
                        "name":"version",
                        "in":"path",
                        "description":"version of the template",
                        "required":true,
                        "type":"string",
                        "pattern":"^\\d+\\.\\d+$"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Ok",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "name":{
                                    "type":"string",
                                    "maxLength":1000
                                },
                                "version":{
                                    "type":"string",
                                    "pattern":"^\\d+\\.\\d+$"
                                },
                                "type":{
                                    "type":"string",
                                    "maxLength":1000
                                },
                                "sub_template":{
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "sub_type": {
                                                "type": "string"
                                            },
                                            "data": {
                                                "type": "string"
                                            },
                                            "render_mode": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "default":{
                                    "type":"boolean"
                                },
                                "attributes":{
                                    "type":"object",
                                    "additionalProperties":true
                                },
                                "sample_context_data":{
                                    "type":"object",
                                    "additionalProperties":true
                                }
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "error":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "tags":[
                    "Create a new version with this one as default version"
                ],
                "summary":"Set default version for the template",
                "description":"",
                "operationId":"post_template_details",
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"name",
                        "in":"path",
                        "description":"name of the template",
                        "required":true,
                        "type":"string"
                    },
                    {
                        "name":"version",
                        "in":"path",
                        "description":"version of the template",
                        "required":true,
                        "type":"string",
                        "pattern":"^\\d+\\.\\d+$"
                    },
                    {
                        "in":"body",
                        "name":"body",
                        "description":"",
                        "required":true,
                        "schema":{
                            "type":"object",
                            "properties":{
                                "default":{
                                    "type":"boolean",
                                    "description":"value should be true, false value is rejected"
                                }
                            }
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Ok",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "name":{
                                    "type":"string",
                                    "maxLength":1000
                                },
                                "version":{
                                    "type":"string",
                                    "pattern":"^\\d+\\.\\d+$"
                                },
                                "default":{
                                    "type":"boolean"
                                },
                                "attributes":{
                                    "type":"object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}